LayoutTests/platform/ios-sim/http/tests/security/dataURL/xss-DENIED-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

This tests that the main frame can't access the contents of an iframe that contains a data: URL loaded page

 PASS: Cross frame access to a data: URL was denied.


--------
Frame: 'aFrame'
--------
Inner iframe.
